Move logic of nested object to a mixin, it will not apply only to Contact
