+++ src/backend/main/py.storage/caliopen_storage/core/mixin.py
@@ -2,6 +2,7 @@
 """Caliop core mixin classes."""
 from __future__ import absolute_import, print_function, unicode_literals
 import logging
+import uuid
 
 from ..exception import NotFound
 
@@ -112,3 +113,60 @@ class MixinCoreRelation(object):
                      (reltype, key, id))
         nested.remove(found)
         return True
+
+
+class MixinCoreNested(object):
+
+    """Mixin class for core nested objects management."""
+
+    def _add_nested(self, column, nested):
+        """Add a nested object to a list."""
+        nested.validate()
+        kls = self._nested.get(column)
+        if not kls:
+            raise Exception('No nested class for {}'.format(column))
+        column = getattr(self.model, column)
+        # Ensure unicity
+        if hasattr(kls, 'uniq_name'):
+            for value in column:
+                uniq = getattr(value, kls.uniq_name)
+                if uniq == getattr(nested, kls.uniq_name):
+                    raise Exception('Unicity conflict for {}'.format(uniq))
+        if hasattr(nested, 'is_primary') and nested.is_primary:
+            for old_primary in column:
+                column.is_primary = False
+        value = nested.to_primitive()
+        pkey = getattr(kls, '_pkey')
+        value[pkey] = uuid.uuid4()
+        log.debug('Will insert nested {} : {}'.format(column, value))
+        column.append(kls(**value))
+        return value
+
+    def _delete_nested(self, column, nested_id):
+        """Delete a nested object with its id from a list."""
+        attr = getattr(self, column)
+        log.debug('Will delete {} with id {}'.format(column, nested_id))
+        found = -1
+        for pos in xrange(0, len(attr)):
+            nested = attr[pos]
+            current_id = str(getattr(nested, nested._pkey))
+            if current_id == nested_id:
+                found = pos
+        if found == -1:
+            log.warn('Nested object {}#{} not found for deletion'.
+                     format(column, nested_id))
+            return None
+        return attr.pop(found)
+
+    @classmethod
+    def create_nested(cls, values, kls):
+        """Create nested objects in store format."""
+        nested = []
+        for param in values:
+            param.validate()
+            attrs = param.to_primitive()
+            # XXX default value not correctly handled
+            pkey = getattr(kls, '_pkey')
+            attrs[pkey] = uuid.uuid4()
+            nested.append(kls(**attrs))
+        return nested
